reflect reviews

+++ core/database-core/src/main/kotlin/com/teamplay/core/database/EntityId.kt
@@ -3,5 +3,5 @@ package com.teamplay.core.database
 import java.io.Serializable
 
 interface EntityId {
-    val id: Int?
+    val id: Long?
 }
